requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.10
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
PyJWT>=2.8.0
httpx>=0.25.0
aiohttp>=3.9.0
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
pinecone-client>=3.0.0
redis>=5.0.0
boto3>=1.28.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
requests>=2.31.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
celery>=5.3.0

